Refactored tokenization: Removed inheritance from tokens. All tokens are now just a ParseToken with a tokenid and a subtype. Added TokenFactory. Specific tokens generators are simplified and inherit from TokenFactory instead of from ParseToken. Token ids and sub types are now unsigned shorts. Tokens are no longer individually allocated on the heap (use std::vector<ParseToken> instead of std::vector<std::unique_ptr<ParseToken>>). Tokens are now stored in the Document. Removed SourceFile as it was now made redundant, use Document instead.

+++ include/jitcat/CommentTokenSubTypes.h
@@ -10,7 +10,7 @@
 namespace jitcat::Tokenizer
 {
 
-	enum class CommentType: int
+	enum class CommentType: unsigned short
 	{
 		SingleLine,
 		Block
